Players : only one class player, that uses symbols based on the symbols available, defined when the game is built with defineSymbols
