+++ game_board/src/fr/lecampusnumerique/cda2025/javaalgo/boardgames/players/IPlayer.java
@@ -1,0 +1,21 @@
+package fr.lecampusnumerique.cda2025.javaalgo.boardgames.players;
+
+import fr.lecampusnumerique.cda2025.javaalgo.boardgames.symbols.Symbol;
+
+public interface IPlayer {
+
+    void setNumber(int number);
+
+    Symbol getSymbol();
+
+    void setSymbol(Symbol symbol);
+
+    boolean isArtificial();
+
+    void setArtificial(boolean artificial);
+
+    String getRepresentation();
+
+    int[] getPlayerMove();
+    void playerTurn();
+}
