+++ game_board/src/fr/lecampusnumerique/cda2025/javaalgo/boardgames/players/Player.java
@@ -2,20 +2,105 @@ package fr.lecampusnumerique.cda2025.javaalgo.boardgames.players;
 
 import fr.lecampusnumerique.cda2025.javaalgo.boardgames.symbols.Symbol;
 
-public interface Player {
+import java.util.Scanner;
 
-    void setNumber(int number);
+public class Player implements IPlayer {
+    private int number;
+    private Symbol symbol;
+    private boolean isArtificial;
 
-    Symbol getSymbol();
+    private void definePlayerSymbol(int playerNumber, Symbol result1, Symbol result2){
+        this.symbol = (playerNumber == 1) ? result1 :result2;
+    }
+    Scanner scanner = new Scanner(System.in);
 
-    void setSymbol(Symbol symbol);
+    public Player(int number, Symbol symbol, boolean isArtificial) {
+        this.number = number;
+        this.symbol = symbol;
+        this.isArtificial = isArtificial;
+    }
 
-    boolean isArtificial();
+    // **
+    // *****
+    // ********* METHODS
+    // *****
+    // **
 
-    void setArtificial(boolean artificial);
+    public int getNumber() {
+        return number;
+    }
 
-    String getRepresentation();
+    public void setNumber(int number) {
+        this.number = number;
+    }
 
-    int[] getPlayerMove();
-    void playerTurn();
+    public Symbol getSymbol() {
+        return this.symbol;
+    }
+
+    public void setSymbol(Symbol symbol) {
+        this.symbol = symbol;
+    }
+
+    public boolean isArtificial() {
+        return isArtificial;
+    }
+
+    public void setArtificial(boolean artificial) {
+        isArtificial = artificial;
+    }
+
+    public String getRepresentation() {
+        return getSymbol().getRepresentation();
+    }
+
+    private int getColumnMove() {
+        boolean running = true;
+        int answer = 0;
+        do {
+            System.out.println("Entrez une colonne :");
+            try {
+                answer = Integer.parseInt(scanner.next());
+                System.out.println("colonne is: " + answer);
+                running = false;
+            } catch (Exception e) {
+                System.out.println("Entry a number!");
+                continue;
+            }
+        } while (running);
+        return answer;
+    }
+
+    private int getRowMove() {
+        boolean running = true;
+        int answer = 0;
+        do {
+            System.out.println("Choisissez une ligne :");
+            try {
+                answer = Integer.parseInt(scanner.next());
+                System.out.println("row is: " + answer);
+                running = false;
+            } catch (Exception e) {
+                System.out.println("Entry a number!");
+                continue;
+            }
+        } while (running);
+        return answer;
+    }
+
+
+    public int[] getPlayerMove() {
+        int[] playerMove = new int[2];
+
+        System.out.println("Get player row");
+        playerMove[0] = getRowMove()-1;
+
+        System.out.println("Get player col");
+        playerMove[1] = getColumnMove()-1;
+
+        return playerMove;
+    }
+
+    public void playerTurn() {
+    }
 }
